WIP restructuring libs
